+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   - Added a Body Mass Index Z-Score calculator based on:
     - http://www.cdc.gov/growthcharts/percentile_data_files.htm
     - http://www.cdc.gov/growthcharts/data/zscore/bmiagerev.csv
+  - Added a Blood Pressure Percentile calculator based on:
+    - http://www.cdc.gov/growthcharts/percentile_data_files.htm
+    - http://www.cdc.gov/growthcharts/data/zscore/statage.csv
 - **Gem Changes**
   - Started testing Ruby 2.3.0-preview1
   - Updated to rails 4.2.5